requirements.txt
# Core dependencies for RAG Test Case Generator
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.2
datasets>=2.14.0
numpy>=1.24.0
pytest>=7.0.0
requests>=2.28.0
requests-mock>=1.11.0

# Retrieval / vector DB / embedding related
faiss-cpu>=1.7.4
langchain>=0.0.350
langchain-huggingface>=0.0.6
langchain-text-splitters>=0.0.1

# Hugging Face tooling
huggingface-hub>=0.14.0
pyarrow>=11.0.0
tiktoken>=0.11.0

# Optional helpers
accelerate>=0.24.0
python-dotenv>=1.0.0
openai>=1.30.0
langchain-openai>=0.1.0
ragas>=0.1.9
